+++ docs/ChangeLog.md
@@ -2,6 +2,7 @@
 
 ### v2.20.x-beta
 
+ * Update google oauth2 to `google-plus` `googlePlus`.
  * Updated / consolidated documentation.
  * Restore auto refresh interval.
  * Add support for promises.